pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sophoset"
version = "0.1.0"
description = "Unified Hugging Face datasets management framework for evaluating vision-language models (VLMs) and large language models (LLMs)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "SophoSet Contributors"}
]
keywords = [
    "datasets",
    "hugging-face",
    "benchmarks",
    "vision-language-models",
    "large-language-models",
    "lmdb",
    "evaluation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "datasets>=2.0.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "transformers>=4.0.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.62.0",
    "requests>=2.26.0",
    "lmdb>=1.0.0",
    "streamlit>=1.0.0",
    "pydantic>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
]
ui = [
    "streamlit>=1.0.0",
    "gradio>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/sophoset"
Documentation = "https://github.com/yourusername/sophoset#readme"
Repository = "https://github.com/yourusername/sophoset.git"
"Bug Tracker" = "https://github.com/yourusername/sophoset/issues"

[tool.setuptools]
packages = ["sophoset", "sophoset.core", "sophoset.text", "sophoset.text.mcq", "sophoset.text.oeq", "sophoset.vision", "sophoset.vision.mcq", "sophoset.vision.oeq", "sophoset.utils"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py40', 'py41', 'py12']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--strict-markers -v"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
